+++ qemu/target-m68k/helper.c
@@ -510,7 +510,7 @@ uint64_t HELPER(macmuls)(CPUM68KState *env, uint32_t op1, uint32_t op2)
     int64_t res;
 
     product = (uint64_t)op1 * op2;
-    res = (product << 24) >> 24;
+    res = ((int64_t)(((uint64_t)product) << 24)) >> 24;
     if (res != product) {
         env->macsr |= MACSR_V;
         if (env->macsr & MACSR_OMC) {
@@ -565,7 +565,7 @@ void HELPER(macsats)(CPUM68KState *env, uint32_t acc)
     int64_t tmp;
     int64_t result;
     tmp = env->macc[acc];
-    result = ((tmp << 16) >> 16);
+    result = ((int64_t)((uint64_t)tmp << 16) >> 16);
     if (result != tmp) {
         env->macsr |= MACSR_V;
     }
@@ -746,12 +746,12 @@ void HELPER(set_mac_extf)(CPUM68KState *env, uint32_t val, uint32_t acc)
     int32_t tmp;
     res = env->macc[acc] & 0xffffffff00ull;
     tmp = (int16_t)(val & 0xff00);
-    res |= ((int64_t)tmp) << 32;
+    res |= ((uint64_t)((int64_t)tmp)) << 32;
     res |= val & 0xff;
     env->macc[acc] = res;
     res = env->macc[acc + 1] & 0xffffffff00ull;
     tmp = (val & 0xff000000);
-    res |= ((int64_t)tmp) << 16;
+    res |= ((uint64_t)((int64_t)tmp)) << 16;
     res |= (val >> 16) & 0xff;
     env->macc[acc + 1] = res;
 }
